Adjust fields type to be compatible with sqlboiler

+++ db/migrations/20170601213757_AddLenghClass.sql
@@ -5,7 +5,7 @@ CREATE TABLE `length_class` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `name` varchar(100) NOT NULL,
   `unit` varchar(4) NOT NULL,
-  `value` decimal(15,8) NOT NULL,
+  `value` double(15,8) NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
